+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Empower cantrip was added
+- Spectral Slime cantrip was added
 
 ### Changed
 
@@ -20,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Spectral Donkey speed on water was fixed
 - Spectral Donkey transparency fix
 - Cantrip delay now works like in the base mod
+- Spectral Armor protection scales with player tier now
 
 ## 0.4.0 - 2026-07-05